Second pass on the types.
